chore: Separate WPF window host and window itself
