+++ src/Uno.UI.Runtime.Skia.Wpf/Extensions/WpfExtensionsRegistrar.cs
@@ -10,11 +10,13 @@ using Uno.Foundation.Extensibility;
 using Uno.Helpers.Theming;
 using Uno.UI.Core.Preview;
 using Uno.UI.Runtime.Skia.Wpf.Extensions.UI.Xaml.Controls;
-using Uno.UI.Runtime.Skia.Wpf.Extensions.UI.Xaml.Input;
 using Uno.UI.Runtime.Skia.Wpf.WPF.Extensions.Helpers.Theming;
+using Uno.UI.Skia;
 using Uno.UI.Skia.Platform;
 using Uno.UI.Xaml.Controls.Extensions;
+using Uno.UI.Xaml.Hosting;
 using Uno.UI.Xaml.Input;
+using Uno.UI.XamlHost.Skia.Wpf;
 using Windows.Graphics.Display;
 using Windows.Networking.Connectivity;
 using Windows.Storage.Pickers;
@@ -35,6 +37,7 @@ internal static class WpfExtensionsRegistrar
 		}
 
 		ApiExtensibility.Register(typeof(Uno.ApplicationModel.Core.ICoreApplicationExtension), o => new CoreApplicationExtension(o));
+		ApiExtensibility.Register<IXamlRootHost>(typeof(Windows.UI.Core.IUnoCorePointerInputSource), o => new WpfCorePointerInputSource(o));
 		ApiExtensibility.Register(typeof(Windows.UI.Core.ICoreWindowExtension), o => new WpfCoreWindowExtension(o));
 		ApiExtensibility.Register(typeof(Windows.UI.ViewManagement.IApplicationViewExtension), o => new WpfApplicationViewExtension(o));
 		ApiExtensibility.Register(typeof(ISystemThemeHelperExtension), o => new WpfSystemThemeHelperExtension(o));
@@ -49,7 +52,6 @@ internal static class WpfExtensionsRegistrar
 		ApiExtensibility.Register(typeof(IClipboardExtension), o => new ClipboardExtensions(o));
 		ApiExtensibility.Register(typeof(IAnalyticsInfoExtension), o => new AnalyticsInfoExtension());
 		ApiExtensibility.Register(typeof(ISystemNavigationManagerPreviewExtension), o => new SystemNavigationManagerPreviewExtension());
-		ApiExtensibility.Register(typeof(IPointerExtension), o => new PointerExtension());
 
 		_registered = true;
 	}
